update (Cient): move Team pages inside About folder.
